// src/webparts/deviation/services/INTERFACES/ILibraryUpload.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
export interface ILibraryUpload {
 
   

    Id: any; 
    Title:any;
    Path:string;
    FileLeafRef: any; 
    FileDirRef: any; 
    FSObjType: any; 
    File_x0020_Type?: any;
    Editor: { Title: any };
    Type:any;

    FileType: any;
    Modified: any;
    ModifiedBy: any;
    Size:any;
    // Id: number; FileLeafRef: string; FileDirRef: string;  FSObjType: number; 
    //             File_x0020_Type?: string;
    //             Modified: string;
    //             Editor: { Title: string };
}